+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "programs/*"
-, "tests"]
+]
 resolver = "2"
 
 [profile.release]
