+++ programs/solana-stablecoin-standard/Cargo.toml
@@ -20,3 +20,17 @@ idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
 [dependencies]
 anchor-lang = "0.31.1"
 anchor-spl = { version = "0.31.1", features = ["token_2022", "associated_token", "metadata"] }
+
+[dev-dependencies]
+litesvm = "0.6.1"
+solana-address = "1.0.0"
+solana-hash = "2.2.1"
+solana-instruction = "2.2.1"
+solana-pubkey = "2.2.1"
+sss-compliance-hook = { path = "../sss-compliance-hook", features = [
+    "cpi",
+] }
+solana-keypair = "2.2.1"
+solana-transaction = "2.2.1"
+solana-signer = "2.2.1"
+spl-token-2022 = { version = "9.0.0", features = ["no-entrypoint"] }
