+++ programs/solana-stablecoin-standard/tests/tests.md
