--- tests/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "tests"
-version = "0.1.0"
-description = "Created with Anchor"
-edition = "2021"
-
-[dependencies]
-anchor-client = "0.31.1"
-anchor-lang = "0.31.1"
-anchor-spl = "0.31.1"
-solana-stablecoin-standard = { version = "0.1.0", path = "../programs/solana-stablecoin-standard" }
-sss-compliance-hook = { version = "0.1.0", path = "../programs/sss-compliance-hook" }
-litesvm = "0.6.1"
-litesvm-token = "0.6.1"
-solana-keypair = "2.2.1"
-solana-sdk = "2.2.1"
-solana-transaction = "2.2.1"
-solana-instruction = "2.2.1"
-solana-pubkey = "2.2.1"
-solana-signer = "2.2.1"
