--- tests/src/lib.rs
@@ -1,4 +1,0 @@
-#[cfg(test)]
-mod test_initialize;
-
-pub mod test_util;
